requirements.txt
annotated-types==0.7.0
anyio==4.8.0
astor==0.8.1
asyncpg==0.30.0
black==24.10.0
certifi==2025.1.31
click==8.1.8
colorama==0.4.6
distro==1.9.0
dotenv==0.9.9
fastapi==0.115.9
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
injector==0.22.0
jiter==0.8.2
mypy-extensions==1.0.0
openai==1.65.3
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
pynest-api==0.3.2
python-dotenv==1.0.1
PyYAML==6.0.2
sniffio==1.3.1
SQLAlchemy==2.0.38
starlette==0.45.3
tqdm==4.67.1
typing_extensions==4.12.2
uvicorn==0.32.1
